requirements.txt
rumps>=0.4.0
pydexcom>=0.2.3
keyring>=24.0.0
google-genai>=1.0.0
truststore>=0.9.0
pyobjc-core>=9.0
pyobjc-framework-Cocoa>=9.0
pyobjc-framework-WebKit>=9.0
